+++ benches/bench.rs
@@ -18,7 +18,7 @@ use rspack_sources::{
 
 use bench_complex_replace_source::benchmark_complex_replace_source;
 use bench_source_map::{
-  benchmark_parse_source_map_from_json, benchmark_source_map_clone,
+  benchmark_parse_source_map_from_json, benchmark_source_map_as_borrowed,
   benchmark_stringify_source_map_to_json,
 };
 
@@ -253,7 +253,7 @@ fn bench_rspack_sources(criterion: &mut Criterion) {
     benchmark_parse_source_map_from_json,
   );
 
-  group.bench_function("source_map_clone", benchmark_source_map_clone);
+  group.bench_function("source_map_as_borrowed", benchmark_source_map_as_borrowed);
 
   group.bench_function(
     "stringify_source_map_to_json",
